requirements.txt
# YouTube API
google-api-python-client==2.108.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Kafka
confluent-kafka==2.3.0

# Data processing (will add more later)
duckdb==1.1.3

# Natural Language Processing
nltk==3.8.1
rake-nltk==1.0.6

# Visualization
streamlit==1.39.0
plotly==5.24.1
wordcloud==1.9.3
matplotlib==3.9.2
pandas==2.2.3

# Prefect for workflow orchestration
prefect==2.14.6
griffe==0.27.2